+++ __tests__/utils.js
@@ -1,0 +1,7 @@
+const utils = require("../utils");
+
+describe("Utilities", () => {
+  it("asdf", () => {
+    expect(utils.getPackageInfo("asdf")).toEqual({ name: "asdf" });
+  });
+});
